feat(trade_executor): add atomic all-or-nothing batch execution mode
Adds atomic all-or-nothing semantics to batch_execute (Issue #793):
portfolio-rebalancing callers need a batch to either fully apply or
fully revert, since a partial fill can leave a worse position than
doing nothing.

Kept `batch_execute(env, trades)` completely unchanged (still
best-effort, still the same signature) rather than adding a required
`atomic` parameter to it directly -- once a Soroban contract function
is deployed, changing its parameter list breaks any caller already
encoding invocations against the old ABI, so "doesn't break existing
callers" has to mean the exact function signature, not just "same
behavior when passed a default". Added a new `batch_execute_atomic(env,
trades, atomic)` entry point instead, with both delegating to a shared
`batch_execute_impl`.

When atomic = true and any trade in the batch fails, the whole call
panics. Soroban reverts every storage effect from a panicking
top-level invocation, so this relies on the host's own transaction-
level atomicity -- including for trades that individually succeeded
earlier in the same batch -- rather than any manual undo bookkeeping.
Verified this actually reverts prior-succeeded state (not just the
failing trade) via a real client call wrapped in catch_unwind, then
asserting the mock portfolio's position count for the successful
trade is back to 0.

Added `atomic_rollback: bool` to `BatchTradeResult` per the issue's
acceptance criteria. Documented plainly in code that this can't
actually reach a real caller on the failure path, since panicking
means the function never returns a value at all -- it's set correctly
for API completeness and is unit-tested in isolation via a small
`mark_atomic_rollback` helper, independent of testing the panic-based
rollback itself.

Also fixes three small, unrelated things discovered while trying to get
a clean local build/lint/format signal for this change:

- contracts/shared/src/multisig.rs: two redundant `as u32` casts on an
  already-u32 value (Vec::len() returns u32 in this soroban-sdk
  version). These, and a manual `impl Default` in
  contracts/common/src/retry_backoff.rs that clippy now flags as
  derivable, were blocking `cargo clippy -p trade_executor --all-targets
  -- -D warnings` from running at all, since both are transitive
  dependencies of trade_executor.
- scripts/build.sh: line 58 sourced
  "$SCRIPT_DIR/scripts/embed_source_hash.sh", double-appending
  "scripts/" since SCRIPT_DIR already points at the scripts/ directory
  containing both files. This breaks the "Reproducible WASM build
  check" CI job for every PR, unrelated to trade_executor -- confirmed
  by reproducing it on a fresh clone of main with none of this PR's
  changes applied.

Did NOT attempt to fix two other pre-existing, unrelated problems
found along the way, both confirmed present on a completely clean
`main` checkout and both far outside this issue's scope:

- `cargo fmt --all -- --check` currently reports 332 diffs across the
  workspace, none in any file this PR touches.
- `cargo build -p trade_executor --target wasm32-unknown-unknown
  --release` fails to link with "duplicate symbol: initialize",
  because contracts/shared is itself built with crate-type = cdylib
  (its own deployable contract exports, including `initialize`)
  and trade_executor depends on it as an ordinary library --
  colliding at link time. This is a structural crate-type issue in
  `shared`, not something touched or introduced by this change, and
  reproduces identically with none of this PR's commits applied.

Both are called out explicitly in the PR description rather than
silently worked around or left for CI to surface without context.

Closes #793

+++ stellar-swipe/contracts/shared/src/multisig.rs
@@ -269,7 +269,7 @@ pub fn approve(env: &Env, caller: &Address, proposal_id: u64) -> Result<u32, Mul
         .persistent()
         .set(&MultisigStorageKey::Proposal(proposal_id), &proposal);
 
-    Ok(count as u32)
+    Ok(count)
 }
 
 /// Return `true` if `signer` has already approved `proposal`.
@@ -312,7 +312,7 @@ pub fn require_can_execute(env: &Env, proposal: &Proposal) -> Result<(), Multisi
     if proposal.status == ProposalStatus::Executed {
         return Err(MultisigError::AlreadyExecuted);
     }
-    if (proposal.approvals.len() as u32) < config.threshold {
+    if proposal.approvals.len() < config.threshold {
         return Err(MultisigError::ThresholdNotMet);
     }
     if env.ledger().timestamp() > proposal.expires_at {
